requirements.txt
click>=8.1.7
Pillow>=10.2.0
PyYAML>=6.0.1
numpy>=1.26.0
